+++ src/v1/guarantees/mustBeUuidData.ts
@@ -1,4 +1,3 @@
-//
 // Copyright (c) 2019-present Ganbaro Digital Ltd
 // All rights reserved.
 //
@@ -31,37 +30,17 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { OnError } from "@ganbarodigital/ts-on-error/V1";
+import { OnError, THROW_THE_ERROR } from "@ganbarodigital/ts-lib-error-reporting/lib/v1";
 
-import { InvalidUuidError } from "../errors";
-import { uuidFromFormatted } from "./formatted";
+import { InvalidUuidError } from "../errors/InvalidUuid";
+import { isUuidData } from "../guards/isUuidData";
 
-describe("Uuid", () => {
-
-    it("accepts a well-formatted UUID string", () => {
-        const inputValue = "123e4567-e89b-12d3-a456-426655440000";
-        const actualValue = uuidFromFormatted(inputValue);
-
-        expect(actualValue).toBeInstanceOf(String);
-    });
-
-    it("accepts an error handler", () => {
-        const onError: OnError<InvalidUuidError> = (reason, description, extra) => {
-            throw extra;
-        };
-        const inputValue = "123e4567-e89b-12d3-a456-426655440000";
-        const actualValue = uuidFromFormatted(inputValue, onError);
-
-        expect(actualValue).toBeInstanceOf(String);
-    });
-
-    it("auto-converts to a string primative", () => {
-        const inputValue = "123e4567-e89b-12d3-a456-426655440000";
-        const expectedValue = "this is a uuid: " + inputValue;
-        const uuid = uuidFromFormatted(inputValue);
-
-        const actualValue = "this is a uuid: " + uuid;
-
-        expect(actualValue).toEqual(expectedValue);
-    });
-});
+/**
+ * throws an error if the given string is not a well-formatted UUID
+ */
+export function mustBeUuidData(input: string, onError: OnError = THROW_THE_ERROR): void {
+    // a string must contain a well-formatted UUID
+    if (!isUuidData(input)) {
+        onError(new InvalidUuidError({public: { invalidInput: input}}));
+    }
+}